Added "generic" UncertaintyElementModel
