+++ SharePointFramework/ProjectWebParts/src/components/RiskMatrix/types.tsx
@@ -3,65 +3,17 @@ import { PageContext } from '@microsoft/sp-page-context'
 import strings from 'ProjectWebPartsStrings'
 import { HTMLProps } from 'react'
 import { IDynamicMatrixProps } from '../DynamicMatrix'
-import { IMatrixElementModel } from '../DynamicMatrix/MatrixCell/MatrixElement/types'
+import { UncertaintyElementModel } from 'types/UncertaintyElementModel'
 
 export interface IRiskMatrixProps
   extends Omit<HTMLProps<HTMLDivElement>, 'size'>,
     Pick<IDynamicMatrixProps, 'size' | 'colorScaleConfig' | 'calloutTemplate'> {
   customConfigUrl?: string
-  items?: RiskElementModel[]
+  items?: UncertaintyElementModel[]
   fullWidth?: boolean
   pageContext?: PageContext
   overrideHeaderLabels?: Record<string, boolean>
   headerLabels?: Record<string, string[]>
-}
-
-export interface IRiskElementItem {
-  Id: number
-  Title: string
-  [key: string]: any
-}
-
-export class RiskElementModel implements IMatrixElementModel {
-  public id: number
-  public title: string
-  public probability: number
-  public consequence: number
-  public probabilityPostAction: number
-  public consequencePostAction: number
-  public action: string
-  public url: string
-  public webId: string
-  public webUrl: string
-  public siteTitle: string
-
-  constructor(
-    public item: IRiskElementItem,
-    probability?: string,
-    consequence?: string,
-    probabilityPostAction?: string,
-    consequencePostAction?: string
-  ) {
-    this.id = item.Id || item.ID
-    this.title = item.Title
-    this.probability = parseInt(probability || item.GtRiskProbability, 10)
-    this.consequence = parseInt(consequence || item.GtRiskConsequence, 10)
-    this.probabilityPostAction = parseInt(
-      probabilityPostAction || item.GtRiskProbabilityPostAction,
-      10
-    )
-    this.consequencePostAction = parseInt(
-      consequencePostAction || item.GtRiskConsequencePostAction,
-      10
-    )
-  }
-
-  public get tooltip() {
-    let tooltip = ''
-    if (this.siteTitle) tooltip += `${this.siteTitle}: `
-    tooltip += this.title
-    return tooltip
-  }
 }
 
 function getHeaderLabel(
