+++ SharePointFramework/ProjectWebParts/src/types/index.ts
@@ -1,1 +1,2 @@
-export { PhaseChecklistData } from './PhaseChecklistData'
+export * from './PhaseChecklistData'
+export * from './UncertaintyElementModel'
